requirements.txt
asgiref==3.4.0
certifi==2021.5.30
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
